requirements.txt
streamlit>=1.25.0
ultralytics
pillow
numpy
opencv-python-headless
